requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
python-multipart
python-dotenv
